refactor(checker): split assignability/property-access/overload-resolution under 2000 LOC (arch-guard)
Pure mechanical code motion (zero behavior change) to bring checker source
files back under the hard 2000-LOC arch-guard limit. Public/pub(crate) APIs
and all call sites are unchanged; extracted methods stay inherent on
CheckerState via additional impl blocks in sibling modules.

- error_reporter/assignability.rs (2091 -> 1722): extract the TS2322 generic
  assignability family (error_type_not_assignable_generic_at /
  _with_anchor) into new sibling error_reporter/assignability_generic.rs.
- types/property_access_type/helpers.rs (2032 -> 1805): extract the
  exported-value-declaration detection cluster
  (symbol_has_exported_value_declaration + check_value_decl_has_export_in_arena)
  into new sibling types/property_access_type/exported_value_decl.rs.
- checkers/call_checker/overload_resolution.rs (2007 -> 17): move the single
  ~1962-line resolve_overloaded_call_with_signatures method into new child
  module overload_resolution/resolve_signatures.rs (1994). The two tiny
  snapshot/rollback_overload_retry_state helpers move into the existing
  call_checker/diagnostics.rs (which already owns speculative-call rollback
  helpers) so the speculation snapshot-rollback caller-file count stays at 4.
- types/utilities/enum_utils.rs (2020 -> 1867): pre-existing LOC violation on
  main, same class; extract the property-readonly / class-name utility cluster
  into new sibling types/utilities/enum_utils_readonly.rs.

Verification: scripts/arch/arch_guard.py reports no LOC failures and no
snapshot-rollback cap failure for these files; cargo clippy -p tsz-checker is
clean (0 warnings, 0 errors).
