+++ crates/tsz-checker/src/types/utilities/mod.rs
@@ -9,5 +9,6 @@ pub(crate) mod core;
 pub(crate) mod cycle_guard;
 pub(crate) mod element_indexable;
 pub(crate) mod enum_utils;
+pub(crate) mod enum_utils_readonly;
 pub(crate) mod overlap_relation_helpers;
 pub(crate) mod return_type;
